+++ test/test_config.yaml
@@ -1,7 +1,7 @@
 # global configuration parameters
 CONFIG: &CONFIG !Config
   version: 0.0.1
-  units: raw &UNITS
+  units: &UNITS raw
   lower_limit: &LIMIT_LOW 0
   upper_limit: &LIMIT_HIGH 20000
   adc_sample_rate: 860
@@ -9,35 +9,36 @@ CONFIG: &CONFIG !Config
   adc_channel: 1
 # define any actions that you want to reuse multiple times in multiple routines
 # ensure you overwrite the nxt table on each usage if needed
-ACTIONS: &ACTIONS
+ACTIONS: &ACTIONS  # actions that chain are done in reverse order, ie last listed first.
+  - !Action &CLEANUP
+      name: CLEANUP
+      nxt:
+        success: None
+  - !Action &RESET_MIN
+      name: RESET_MIN
+      params:
+        low: *LIMIT_LOW
+        high: *LIMIT_HIGH
+      nxt:
+        success: *CLEANUP
   - !Action &RESET_MAX
     name: RESET_MAX
     params:
-      pos:
-        - 1
-        - 2
-        - 3
+      low: *LIMIT_LOW
+      high: *LIMIT_HIGH
     nxt:
-      True: RESET_MIN
-      failed: TEST
-      False: RESET_MAX  # loop back state
+      success: *RESET_MIN
+      failure: *CLEANUP
 # begin routine definitions
 ROUTINES: &ROUTINES
   - !Routine
     name: reset_min
-    units: &UNITS
+    units: *UNITS
     actions:
-      - !Action
-          name: RESET_MIN
-          params:
-            low: *LIMIT_LOW
-            high: *LIMIT_HIGH
-          nxt:
-            True: CLEANUP
-            False: RESET_MIN  # loop back state -- completely useless usually
+      - *RESET_MIN
   - !Routine
     name: reset_max
-    units: raw
+    units: *UNITS
     output: stdout
     actions:
       - *RESET_MAX
